script/requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
billiard==3.3.0.23
celery==3.1.23
kombu==3.0.35
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
click==6.7
python-dateutil==2.6.0
elasticsearch==5.4.0
urllib3==1.21.1
certifi